Update Pom.xml, Security configuration, Admin controller, and User controller
The version of Spring Boot and Spring Security have been updated in pom.xml, the description of the application has been shortened to reflect the nature of the project.

The SecurityConfiguration file was altered to include new options when configuring HttpSecurity.

The Admin controller now has new methods for managing movies, adding and deleting movies.

The User controller now includes functionality related to adding movies to a cart and making purchases.