Refactor create movie IDs with sequences
In order to keep better track of movie IDs created in the application, we've utilized a sequence generator for the creation of unique IDs. Movie objects now have an ID field included in their SQL INSERT statements. Additionally, a new `deleteMovie` endpoint has been added to the `UserController` with support for two `@PostMapping` request mappings. Finally, we've removed the delete functionality from the admin home page, instead relying solely on the `deleteMovie` endpoint.